add debug output of received events
the received events are reencoded as json. this doesn't catch erroneous input which breaks the json.Decoder. logging these errors is at first glance a bit trickier as the incoming stream is read blockwise by the decoder. this leads to chunked json when logging the input using a TeeReader. if the need for such a debug output is great, it's maybe added later.